pyproject.toml
[project]
name = "exparso"
version = "0.0.2"
description = "Analyzing and parsing documents"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = [ "langchain", "openai", "pdfplumber", "pillow", "openpyxl", "pdf"]
dependencies = [
    "langchain>=0.3.14",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.4",
    "pillow>=10.4.0",
]

[tool.uv]
default-groups = []

[dependency-groups]
dev = [
    "black>=24.10.0",
    "flake8>=7.1.1",
    "flake8-pyproject>=1.2.3",
    "isort>=5.13.2",
    "pip-licenses>=5.0.0",
    "pre-commit>=4.0.1",
    "pylance>=0.18.2",
]
test = [
    "pytest>=8.3.3",
    "pytest-dotenv>=0.5.2",
    "pytest>=8.3.3",
    "pytest-dotenv>=0.5.2",
    "openai>=1.51.2",
    "langchain-openai>=0.3.0",
    "anthropic[vertex]>=0.36.0",
    "google-cloud-aiplatform>=1.70.0",
    "langchain-google-vertexai<=2.0.10",
]
eval = [
    "google-cloud-aiplatform>=1.70.0",
    "langchain-google-vertexai<=2.0.10",
    "azure-ai-documentintelligence==1.0.0",
    "azure-identity>=1.19.0",
    "docling==2.15.1",
    "pymupdf4llm==0.0.17",
    "google-cloud-storage>=2.18.2",
    "matplotlib>=3.9.2",
    "pandas>=2.2.3",
    "tqdm>=4.66.5",
    "langfuse>=2.57.0",
    "pydantic-settings>=2.7.1",
]
deploy = [
    "twine>=5.1.1",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.44",  
    "pymdown-extensions>=10.11.2",    
]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.venv
  | \.git
)
'''

[tool.flake8]
exclude = [".venv", ".git", "__pycache__"]
ignore = ["E203", "W503", "W504", "E701", "E704"]
max-line-length = 120

[tool.isort]
profile = 'black'
line_length = 120

[tool.pytest.ini_options]
pythonpath = "parse_docs"
testpaths = "tests"
addopts = "-s"
env_files = [".env"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = ["tests", "eval", "docs", "uv.lock", "*.yml", ".*", "*.md"]

[tool.hatch.build.targets.wheel]
exclude = ["tests", "eval", "docs", "uv.lock", "*.yml", ".*", "*.md"]
